GET, PUT, DELETE now check user permission.

+++ src/main/java/DeleteHandler.java
@@ -1,4 +1,5 @@
 import com.sun.net.httpserver.HttpExchange;
+import utils.RestrictedAccessException;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -11,12 +12,12 @@ import java.nio.file.Paths;
 public class DeleteHandler extends AbstractHttpHandler {
     @Override
     public void handle(HttpExchange t) throws IOException {
+
         String filename = t.getRequestURI().toString();
         try {
+            checkPermission(t);
+
             String fullFileName = String.format("%s%s", Server.FILES_DIR, filename);
-
-
-
             Files.delete(Paths.get(fullFileName));
             FileCacheService.getInstance().removeFile(fullFileName);
             System.out.printf("Succesfully deleted %s%n", fullFileName);
@@ -25,9 +26,11 @@ public class DeleteHandler extends AbstractHttpHandler {
             System.err.format(String.format("%%s: no such file or directory%%n"), filename);
         } catch (IOException x) {
             System.err.println(x);
+        } catch (RestrictedAccessException e) {
+            System.out.println("Access restricted!");
+            sendResponseAndClose(403, "Access resricted!", t);
+            return;
         }
-
-
 
         sendResponseAndClose(300, "Serverside error, sorry!", t);
     }
